% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{check_state}
\alias{check_state}
\title{Check state column}
\usage{
check_state(state_name)
}
\arguments{
\item{state_name}{Two letter abbreviation of state name or full name}
}
\description{
State should be either a two letter abbreviation or full state name. Check to make sure this is true.
}
\keyword{internal}
