% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{federal_poverty_guidelines}
\alias{federal_poverty_guidelines}
\title{Federal poverty guideline amounts}
\usage{
federal_poverty_guidelines(year, geography, household_size)
}
\arguments{
\item{year}{Year to use to calculate threshold, as numeric. Data for 2016 to 2021 is available.}

\item{geography}{There are different poverty guidelines for the contiguous 48 states, Alaska, and Hawaii.
Use 'Contiguous US' for the contiguous 48 states, 'Alaska' for Alaska, or 'Hawaii' for Hawaii.}

\item{household_size}{Household sizes have different poverty threshold. Specify household size with a
whole number, as a numeric data type, representing the household size.}
}
\value{
A single number representing the poverty threshold for the given year, geographic area, and family size.
}
\description{
Calculate the federal poverty guidelines from the year, geographic area, and household size.
Poverty guidelines are used to calculate eligibility for public benefits. These are not the poverty
thresholds, which are the statistics used to calculate official poverty levels.
}
\section{Source}{

HHS Poverty Guidelines: \url{https://aspe.hhs.gov/topics/poverty-economic-mobility/poverty-guidelines}
}

\examples{
federal_poverty_guidelines('2021', 'Contiguous US', 4)

## You can calculate multiple household sizes at once
federal_poverty_guidelines('2021', 'Hawaii', 1:5)

## SNAP requires an eligibility threshold that is 130\% of the federal poverty line
federal_poverty_guidelines('2021', 'Contiguous US', 3) * 1.3

\dontrun{
## This does not work, however.
federal_poverty_guidelines('2021', c('Contiguous US', 'Hawaii'), 3)
}

}
