% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{snap_calculate_net_income}
\alias{snap_calculate_net_income}
\title{Calculate net income}
\usage{
snap_calculate_net_income(
  net_income_before_shelter,
  shelter_expenses,
  use_homeless_shelter_deduction,
  year
)
}
\value{
Net income value.
}
\description{
Net income is calculated by taking earned and unearned income, subtracting deductions, and then
calculating and subtracting the excess shelter deduction. Steps to calculate net income are in 7 CFR 273.10(e)(1)(i).
}
\keyword{internal}
