% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{snap_determine_eligibility}
\alias{snap_determine_eligibility}
\title{Calculate Eligibility}
\usage{
snap_determine_eligibility(
  net_income,
  monthly_earned_income,
  monthly_unearned_income,
  net_income_limit,
  gross_income_limit,
  elderly_disabled_household_member
)
}
\value{
Boolean representing whether household is eligible for benefits
}
\description{
Calculate Eligibility
}
\section{Eligibility Criteria}{

Disables or elderly compare their net income with net income limit (7 CFR 273.10(e)(2)(i)(A)).
Non-elderly compare their net and gross incomes (7 CFR 273.10(e)(2)(i)(A)). to the limits.
}

\keyword{internal}
