% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{snap_income_deduction}
\alias{snap_income_deduction}
\title{SNAP income deductions}
\usage{
snap_income_deduction(year, deduction)
}
\arguments{
\item{year}{Numeric. Year calculating benefits for.}

\item{deduction}{String. Income deduction. One of 'standard', 'earned_income', 'excess_medical',
'dependent_care', 'child_support', 'homeless_shelter', 'excess_shelter'}
}
\value{
A single number representing a value used to compute the income deduction.
}
\description{
Values that control income deductions. These values need to be checked yearly at 7 CFR 273.9(d).
}
\keyword{internal}
