% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{snap_maximum_benefit_amounts}
\alias{snap_maximum_benefit_amounts}
\title{Maximum benefit values for SNAP}
\usage{
snap_maximum_benefit_amounts(year, region)
}
\arguments{
\item{year}{Fiscal year to use to calculate threshold, as numeric. Fiscal year starts Oct. 1.}

\item{region}{Region to calculate maximum benefits for. Must be one of :
'Contiguous US', 'Guam', 'Virgin Islands', 'Hawaii', 'Alaska Urban', 'Alaska Rural 1', 'Alaska Rural 2'.}
}
\description{
Maximum SNAP benefits depend on year, region, and household size.
}
\keyword{internal}
