% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{snap_maximum_benefits}
\alias{snap_maximum_benefits}
\title{Maximum SNAP Benefits}
\usage{
snap_maximum_benefits(year, region, household_size)
}
\arguments{
\item{year}{Whole number representing the fiscal year. Fiscal years start Oct. 1st.}

\item{region}{Region for calculating max benefits. One of:
'Contiguous US', 'Guam', 'Virgin Islands', 'Hawaii', 'Alaska Urban', 'Alaska Rural 1', 'Alaska Rural 2'}

\item{household_size}{Whole number represent household size.}
}
\description{
Calculates the maximum SNAP benefits for a given year, region, and household size.
Maximum SNAP benefits are based on Thrift Food Plan (TFP) amounts. 7 CFR 273.10(e)(1)(i).
}
\section{Source}{

TFP amounts: \url{https://fns-prod.azureedge.net/sites/default/files/resource-files/2022-SNAP-COLA-\%20Maximum-Allotments.pdf}
}

