% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{snap_net_income_prior_shelter}
\alias{snap_net_income_prior_shelter}
\title{Calculate net income prior to shelter deduction}
\usage{
snap_net_income_prior_shelter(
  year,
  net_income_limit,
  monthly_earned_income,
  monthly_unearned_income,
  elderly_disabled_household_member,
  excess_medical_deduction,
  dependent_care_deduction,
  child_support_deduction
)
}
\description{
Overall net income calculations are in 7 CFR 273.10(e)(1)
}
\keyword{internal}
