% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{snap_state_gross_income_limits}
\alias{snap_state_gross_income_limits}
\title{State gross income limits for SNAP}
\usage{
snap_state_gross_income_limits(year, state)
}
\arguments{
\item{year}{Year to use to calculate threshold, as numeric. Data for 2021 is available.}

\item{state}{Two letter abbreviation of the state, capitalized. Use 'DC' for Washington, DC;
'GU' for Guam, and 'VI' for the US Virgin Islands.}
}
\value{
A single number representing the state's gross income limit, based on the percentage of the
federal poverty guidelines. For example, 2 represents 200\% of the federal poverty guidelines,
while 1.85 represents 185\% of the federal poverty guidelines.
}
\description{
For SNAP benefits, states can elect broad-based categorical eligibility. With this election, states
can increase the gross income limit from 130\% of the poverty guideline to a higher value.
This function returns a state's gross income limit under broad-based categorical eligibility.
An NA value is returned for states that have not elected broad-based
categorical eligibility.
}
\section{Source}{

FNS Broad-Based Categorical Eligibility Chart: \url{https://fns-prod.azureedge.net/sites/default/files/resource-files/BBCE\%20States\%20Chart\%20(July\%202021).pdf}
}

\examples{
snap_state_gross_income_limits(2021, 'NC')

}
